# backend/openapidoc.yaml
openapi: 3.0.0
info:
  version: 0.0.1
  title: Places server API
  description: Simple REST API showing favorite places of users
  contact:
    name: John Doe
    email: john.doe@dontuse.com
servers:
  - url: https://jkserver.azurewebsites.net/api
    description: Production server
  - url: http://localhost:3001/api
    description: Test server running locally
security:
  - bearerAuth: [] 
paths:
...
  /users/{userID}:
    get:
      operationId: getUser
      summary: Fetch the user of favorite places app
      tags:
      - users
      description: Returns the data relating to certain user of favorite places app
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: number
          example: 1
      responses:
        '200':
          description: User data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/users"    
        '404':
          description: Something went bad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"    
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    users:
      type: object
      properties:
        ID:
          type: number
        Name:
          type: string
          description: Name of the registered users
          example: Radu
        Password:
          type: string
          description: Hashed password of the user
          example: $2b$10$Td7EwsHrdB/UV1OslDm/kOZMxJ/t3tHrKM2NxpnhozBtK1Losbk.9
    error:
      type: object
      properties:
        text:
          type: string
          description: The error description
          example: Unknown column a in where clause
        sql:
          type: string
          description: Origin of SQL error
          example: SELECT * FROM places WHERE ID=x"
        fatal:
          type: boolean
          description: A boolean value indicating whether the connection remains valid
          example: false
        errno:
          type: number
          description: The rrror number
          example: 1000
        sqlState:
          type: number
          description: The SQL state code
          example: 40000
        code:
          type: string
          description: The error code
          example: ER_BAD_FIELD_ERROR